# egov-financials/swagger.yaml
...
swagger: '2.0'
...
info:
  version: "1.0.0"
  title: Financial Management System APIs
  host: phoenix-qa.egovernments.org
  description: All APIs available in the system are defined here.
  contact:
    name: Manikanta PT
schemes:
  - https
  - http
basePath: '/v1/egf'
paths:
    /banks:
      get:
        summary: Get the list of banks
        description: banks
        tags: 
          - Masters
          - Bank
        parameters:
          - name: requestInfo.action
            type: string
            in: query
            description: action
          - name: requestInfo.did
            type: string
            in: query
            description: did
          - name: requestInfo.msgId
            type: string
            in: query
            description: msg id
          - name: requestInfo.requesterId
            type: string
            in: query
            description: requester id
          - name: requestInfo.authToken
            type: string
            in: query
            description: auth token
          - name: requestInfo.apiId
            type: string
            in: query
            description: api id
          - name: requestInfo.ver
            type: string
            in: query
            description: ver
          - name: requestInfo.ts
            type: string
            format: date
            in: query
            description: ts
          - name: requestInfo.key
            type: string
            in: query
            description: key
          - name: requestInfo.tenantId
            type: string
            in: query
            description: tenant id
          - name: bank.id
            type: integer
            format: int64
            in: query
            description: Unique Identifier of the Bank
          - name: bank.code
            type: string
            in: query
            description: code of the Bank
            required: true
          - name: bank.name
            type: string
            in: query
            description: name of the Bank
            required: true
          - name: bank.description
            type: string
            in: query
            description: description of the Bank
          - name: bank.active
            type: boolean
            in: query
            description: Whether Bank is Active or not. If the value is TRUE, then Bank is active,If the value is FALSE then Bank is inactive
            required: true
          - name: bank.type
            type: string
            in: query
            description: type of the Bank
          - name: page.totalResults
            type: integer
            format: int32
            in: query
            description: total results of the Search Results
          - name: page.totalPages
            type: integer
            format: int32
            in: query
            description: total pages of the Search Results
          - name: page.pageSize
            type: integer
            format: int32
            in: query
            description: page size of the Search Results
          - name: page.currentPage
            type: integer
            format: int32
            in: query
            description: current page of the Search Results
        responses:
          200:
            description: Bank retrieved Successfully
            schema:
              $ref: '#/definitions/BankResponse'
          400:
            description: Invalid Input
            schema:
              $ref: '#/definitions/ErrorResponse'
          404:
            description: Bank Not Found
      post:
        summary: Create  new  banks
        description: Create  new banks
        tags: 
          - Masters
          - Bank
        parameters:
          - name: bankRequest
            in: body
            description: Create  new 
            schema:
              $ref: '#/definitions/BankRequest'
        responses:
          201:
            description: Bank created Successfully
            schema:
              $ref: '#/definitions/BankResponse'
          400:
            description: Invalid Input
            schema:
              $ref: '#/definitions/ErrorResponse'
          404:
            description: Bank Not Found
      put:
        summary: Update any of the banks
        description: Update any of the banks
        tags: 
          - Masters
          - Bank
        parameters:
          - name: bankRequest
            in: body
            description: common request_info
            schema:
              $ref: '#/definitions/BankRequest'
        responses:
          200:
            description: Bank update Successfully
            schema:
              $ref: '#/definitions/BankResponse'
          400:
            description: Invalid Input
            schema:
              $ref: '#/definitions/ErrorResponse'
          404:
            description: Bank Not Found
...
definitions:
    Bank:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique Identifier of the Bank
        code:
          type: string
          description: code of the Bank
          maxLength: 50
        name:
          type: string
          description: name of the Bank
          maxLength: 100
        description:
          type: string
          description: description of the Bank
          maxLength: 250
        active:
          type: boolean
          description: Whether Bank is Active or not. If the value is TRUE, then Bank is active,If the value is FALSE then Bank is inactive
        type:
          type: string
          description: type of the Bank
          maxLength: 50
    BankRequest:
      properties:
        requestInfo:
          $ref: '#/definitions/RequestInfo'
        banks:
          type:  array
          items:
            $ref: '#/definitions/Bank'
    BankResponse:
      properties:
        responseInfo:
          $ref: '#/definitions/ResponseInfo'
        banks:
          type:  array
          items:
            $ref: '#/definitions/Bank'
        page:
          $ref: '#/definitions/Page'
...
    ResponseInfo:
     type: object
     properties:
      resMsgId:
       type: string
       description: res msg id of the ResponseInfo
      status:
       type: string
       description: status of the ResponseInfo
      apiId:
       type: string
       description: api id of the ResponseInfo
      ver:
       type: string
       description: ver of the ResponseInfo
      ts:
       type: string
       format: date
       description: ts of the ResponseInfo
      key:
       type: string
       description: key of the ResponseInfo
      tenantId:
       type: string
       description: tenant id of the ResponseInfo
    RequestInfo:
     type: object
     properties:
      action:
       type: string
       description: action of the RequestInfo
      did:
       type: string
       description: did of the RequestInfo
      msgId:
       type: string
       description: msg id of the RequestInfo
      requesterId:
       type: string
       description: requester id of the RequestInfo
      authToken:
       type: string
       description: auth token of the RequestInfo
      apiId:
       type: string
       description: api id of the RequestInfo
      ver:
       type: string
       description: ver of the RequestInfo
      ts:
       type: string
       format: date
       description: ts of the RequestInfo
      key:
       type: string
       description: key of the RequestInfo
      tenantId:
       type: string
       description: tenant id of the RequestInfo
    ErrorResponse:
     type: object
     properties:
      responseInfo:
        $ref: "#/definitions/ResponseInfo"
      error:
        $ref: "#/definitions/Error"
    Error:
     type: object
     properties:
      code:
       type: integer
       format: int32
       description: Code of the Error
      message:
       type: string
       description: Message of the Error
      description:
       type: string
       description: Description of the Error
      filelds:
       type: array
       description: Filelds of the Error
       items:
         $ref: "#/definitions/FieldError"
    FieldError:
     type: object
     properties:
      code:
       type: string
       description: Code of the FieldError
      description:
       type: string
       description: Description of the FieldError
    Page:	              
     type: object
     properties:
      pageSize:
       type: integer
       format: int32
       description: Page Size